+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to the "vscode-phpunit" extension will be documented in this
 Check [Keep a Changelog](http://keepachangelog.com/) for 
 recommendations on how to structure this file.
 
+## 0.2.17
+- fix bug can't find vendor/bin/phpunit
+- filesystem add two methods [tmpfile, getContent]
+
 ## 0.2.16
 - output add two empty line